Add files via upload
This repository features an in-depth analysis of Environmental Protection Agency (EPA) violators across different U.S. states. The analysis was conducted using Excel for data preparation and Power BI for visualization and insights. The goal of this project was to identify trends, patterns, and hotspots of EPA violations, providing a clear understanding of environmental compliance across the country. The findings serve as a valuable resource for policymakers, environmentalists, and researchers interested in environmental protection and regulatory enforcement.